requirements_dev.txt
Django==3.1.4
djangorestframework==3.12.2
coreapi==2.3.3
isort==5.6.4
black==20.8b1
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
Faker==5.0.1
factory-boy==3.1.0
PyYAML==5.3.1
requests==2.25.0

